+++ src/router/router.jsx
@@ -2,7 +2,7 @@ import { createBrowserRouter } from 'react-router-dom';
 import Layout from '../components/Layout/SideBarLayout';
 import MyPage from '../pages/MyPage/MyPage';
 import Dashboard from '../pages/Dashboard/DashboardPage';
-import MyDashboardPage from '../pages/MyDashboardPage';
+import MyDashboardPage from '../pages/Dashboard/MyDashboardPage';
 
 const router = createBrowserRouter([
   {
